refactor: レシピ一覧から不要になったpropsを削除
